TAA-872: Fix error handling on pipes.
Check correct error status so we report
fail of the build command not success of the pipe.

Fall through on failed unit tests to coverage
because running subset of unit tests will fail because
some tests are simply not executed and marked failed.

Misc: ensure disk is clean before clone.