TAA-381: Move OpenBSW builds from jenkins paths
Jenkins folders and jobs can contain white spaces, but this can
cause issues when building from within the workspace. The paths
may not be properly escaped, leading to conflicts with build tools.

To avoid these issues, it's recommended to move builds to a safe
directory that doesn't contain white spaces. This ensures that the
build paths don't conflict with build tools.

By moving builds to a safe directory, you also ensure that the
build directory is easily accessible if the user retains the
instance and connects as the container name 'builder'. The build
directory will be located under /home/builder/bsw-builds, rather
than the Jenkins workspace path, e.g.
/home/jenkins/agent/workspace/OpenBSW/Builds/BSW Builder

This simplifies the build directory structure and makes it easier
to manage.

The workspace path is still required for Jenkins to support coverage
reports etc, so all reports are copied back to workspace.