TAA-922: MTK Connect support for non-sequential ports
Ports do not have to be sequential, also the host ip address can
change.  As such this code now creates a dynamic list of the host
ip and ports, and when the testbench device is created, it accesses
that list converted to an array.

This future proofs MTK Connect for such features as cloud android
orchestration.