TAA-929: Jenkins credentials incorrect.
The bas64 value is the base64 value with no quotes whereas
the secret is stored with quotes. It not match k8s and thus
server and uploader abfs_license fails.

Use kubectl get secrets to retrieve the real secret and
use that instead.

Note: this is just a temporary workaround while
the whole 'lost in translation' for the ABFS license secret is
addressed. See where we store the secret in Jenkins so any
fix can be verified quickly.